Fixed modal rendering for participants page as organizer Created new util function addTeamCodeToProjectAndFilterNoTeam to reduced repeated code on ProjectTable component and Projects by challenge page and prevent direct edit of the store
